moved context pool to cli utils
